refactor: separate Reminder as component
